# src/main/resources/api-specs.yaml
openapi: 3.0.0
info:
  title: Library REST API
  description: Example about REST in Spring boot
  version: 1.0.1
servers:
  - url: http://localhost:8080
    description: Optional server description, e.g. Internal staging server for testing
paths:
  /books:
    post:
      summary: Register a new book in the library.
      description: Register a new book in the library.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        '201':    # status code
          description: The book gets registered in the library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOutput"
          headers:
            Location:
              description: The URI of the registered book
              schema:
                type: string
    get:
      summary: Get an array of book that match the criteria.
      description: Get an array of book that match the criteria.
      parameters:
        - in: query
          name: title
          required: false
          description: Title
          schema:
            type: string
        - in: query
          name: author
          required: false
          description: Author.
          schema:
            type: string
        - in: query
          name: publisher
          required: false
          description: Publisher.
          schema:
            type: string
        - in: query
          name: yearPublished
          required: false
          description: Year published
          schema:
            type: number
            format: int32
      responses:
        '201':    # status code
          description: The book gets registered in the library.
          headers:
            Location:
              description: The URI of the registered book
              schema:
                type: string
  /books/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: number
          format: int32
        required: true
        description: Book id
    get:
      summary: Get book by ID.
      description: Get book by ID.
      responses:
        '200':    # status code
          description: A JSON representation of a book.
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/BookOutput'
    put:
      summary: Update an existing book.
      description: Update an existing book.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        '201':    # status code
          description: The book gets registered in the library.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookOutput"
          headers:
            Location:
              description: The URI of the registered book
              schema:
                type: string
        '204':    # status code
          description: Book info gets updated.
    delete:
      summary: Delete an existing book.
      description: Delete an existing book.
      responses:
        '204':    # status code
          description: The book gets deleted from the repository.
components:
  schemas:
    BookBase:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        publisher:
          type: string
        yearPublished:
          type: number
          format: int32
      required:
        - title
        - author
        - publisher
    BookInput:
      allOf: 
        - $ref: '#/components/schemas/BookBase'
      type: object
    BookOutput:
      allOf: 
        - $ref: '#/components/schemas/BookBase'
      type: object
      properties:
        id:
          type: number
          format: int32